+++ src/GhostResource.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Scriptotek\Alma;
+
+interface GhostResource {
+    public function init($data = null);
+    public function getData();
+}