+++ src/GhostResourceList.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Scriptotek\Alma;
+
+interface GhostResourceList {
+    public function init();
+    public function get();
+}