+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `InvalidQueryException` exception.
 - Added methods `getBarcodes()` and `getUniversityIds()` to `UserIdentifiers` to get all active values.
   These are accessible on the `User` object as `$user->barcodes` and `$user->universityIds`.
+- Added method `$client->items->fromBarcode(...)`
 
 ### Changed